# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Demo Application API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for our in class demo application"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Award": {
        "type": "object",
        "required": [
          "name",
          "year"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Firestore auto-generated document ID",
            "example": "qZxE9HkR2Ljs8Df23GmP"
          },
          "name": {
            "type": "string",
            "description": "Name of the award",
            "example": "Ballon d'Or"
          },
          "year": {
            "type": "number",
            "description": "Year the award was received",
            "example": 2017
          }
        }
      },
      "Club": {
        "type": "object",
        "required": [
          "name",
          "seasons"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier for a club",
            "example": "PZyY4KR2ljsdlFh92HjR"
          },
          "name": {
            "type": "string",
            "description": "The club name",
            "example": "Real Madrid"
          },
          "seasons": {
            "type": "string",
            "description": "Seasons played for this club (e.g., 2009–2018)",
            "example": "2009–2018"
          },
          "goals": {
            "type": "number",
            "description": "Total goals scored for this club",
            "example": 450
          },
          "assists": {
            "type": "number",
            "description": "Total assists made for this club",
            "example": 131
          }
        }
      },
      "Match": {
        "type": "object",
        "required": [
          "opponent",
          "date"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier for a match",
            "example": "G5gRYX36bllXVjaY6ESh"
          },
          "opponent": {
            "type": "string",
            "description": "The opponent team name",
            "example": "Barcelona"
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "The match date",
            "example": "2025-05-01"
          },
          "goals": {
            "type": "number",
            "description": "Number of goals scored",
            "example": 2
          },
          "assists": {
            "type": "number",
            "description": "Number of assists made",
            "example": 1
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/awards": {
      "get": {
        "summary": "Retrieves a list of awards",
        "tags": [
          "Awards"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of Awards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Award"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new award",
        "tags": [
          "Awards"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "year"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Ballon d'Or"
                  },
                  "year": {
                    "type": "number",
                    "example": 2018
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Award created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Award"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/awards/{id}": {
      "get": {
        "summary": "Retrieves an award by ID",
        "tags": [
          "Awards"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Award document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Award retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Award"
                }
              }
            }
          },
          "404": {
            "description": "Award not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing award record",
        "tags": [
          "Awards"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Award document ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Award"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Award updated successfully"
          },
          "404": {
            "description": "Award not found"
          }
        }
      },
      "delete": {
        "summary": "Delete an award record",
        "tags": [
          "Awards"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Award document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Award deleted successfully"
          },
          "404": {
            "description": "Award not found"
          }
        }
      }
    },
    "/clubs": {
      "get": {
        "summary": "Retrieves a list of clubs",
        "tags": [
          "Clubs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of Clubs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Club"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new club record",
        "tags": [
          "Clubs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "season",
                  "goals",
                  "assists"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Manchester United"
                  },
                  "seasons": {
                    "type": "string",
                    "example": "2003-2009"
                  },
                  "goals": {
                    "type": "number",
                    "example": 145
                  },
                  "assists": {
                    "type": "number",
                    "example": 37
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Club created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Club"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/clubs/{id}": {
      "get": {
        "summary": "Retrieves a club by ID",
        "tags": [
          "Clubs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Club document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Club retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Club"
                }
              }
            }
          },
          "404": {
            "description": "Club not found"
          }
        }
      },
      "put": {
        "summary": "Update an existing club record",
        "tags": [
          "Clubs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Club document ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Club"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Club updated successfully"
          },
          "404": {
            "description": "Club not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a club record",
        "tags": [
          "Clubs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Club document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Club deleted successfully"
          },
          "404": {
            "description": "Club not found"
          }
        }
      }
    },
    "/matches": {
      "get": {
        "summary": "Retrieves a list of matches",
        "tags": [
          "Matches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of Matches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Match"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new match record",
        "tags": [
          "Matches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "opponent",
                  "date",
                  "goals",
                  "assists"
                ],
                "properties": {
                  "opponent": {
                    "type": "string",
                    "example": "Barcelona"
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-04-18"
                  },
                  "goals": {
                    "type": "number",
                    "example": 2
                  },
                  "assists": {
                    "type": "number",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Match created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Match"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        }
      }
    },
    "/matches/{id}": {
      "get": {
        "summary": "Retrieves a list of matches",
        "tags": [
          "Matches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of Matches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Match"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an existing match record",
        "tags": [
          "Matches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Match document ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Match"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Match updated successfully"
          },
          "404": {
            "description": "Match not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a match record",
        "tags": [
          "Matches"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Match document ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Match deleted successfully"
          },
          "404": {
            "description": "Match not found"
          }
        }
      }
    }
  },
  "tags": []
}